requirements.txt
# Основные библиотеки
aiogram>=3.0.0  # Асинхронная библиотека для создания Telegram ботов
aiosqlite>=0.19.0  # Асинхронная библиотека для работы с SQLite
plotly>=5.18.0  # Библиотека для создания интерактивных графиков
openai>=1.6.1  # Библиотека для работы с ChatGPT API

# Дополнительные утилиты
python-dotenv>=1.0.0  # Загрузка переменных окружения из .env файла (опционально)
pandas>=2.1.4  # Библиотека для обработки данных (используется с plotly)
pytz>=2024.1  # Библиотека для работы с часовыми поясами 